Updated From Short

+++ Plugins/BCClient/Source/BCClientPlugin/Private/BrainCloudRelayComms.h
@@ -52,7 +52,7 @@ public:
 	BrainCloudRelayComms(BrainCloudClient *client);
 	~BrainCloudRelayComms();
 
-	int64 ping();
+	int32 ping();
 	uint8 netId();
 
 	void connect(BCRelayConnectionType in_connectionType, const FString &in_connectOptionsJson, IServerCallback *callback);
@@ -95,7 +95,7 @@ private:
 	void setupWebSocket(const FString &in_url);
 	void sendPing();
 	TArray<uint8> appendHeaderData(uint8 in_controlByte);
-	TArray<uint8> fromShortBE(short number);
+	TArray<uint8> fromShortBE(int16 number);
 
 	BrainCloudClient *m_client;
 	IServerCallback *m_appCallback;
@@ -111,9 +111,9 @@ private:
 	bool m_bIsConnected;
 	float m_pingInterval;
 	float m_timeSinceLastPingRequest;
-	int64 m_lastNowMS;
-	int64 m_sentPing;
-	int64 m_ping;
+	double m_lastNowMS;
+	double m_sentPing;
+	int16 m_ping;
 	short m_netId = -1;
 
 	const int SIZE_OF_LENGTH_PREFIX_BYTE_ARRAY = 2;
